+++ CHANGELOG.md
@@ -3,13 +3,14 @@
 A Work in Progress (WIP).
 Not yet ready for use.
 
-## v0.1.1 - TBD WIP
+## v0.2.0 - 2023-08-07 WIP
 - Minor credo refactors.
 - Brands: Dankort, GPN, LankaPay, Troy, UkrCard, Verve.
+  - Brands are now complete.
 - In-line document tests.
 - Correct major error of `card_brand/1` passing CodePoints to `card_brand/2`.
 - Identify PAN starting just 65 as Discover, vs. RuPay or Troy.
-- Generating documentation
+- Generate documentation.
 
 ## v0.1.0 - 2023-08-05 WIP
 - Add guards to `card_brand`.